+++ src/repository/mod.rs
@@ -1,8 +1,12 @@
+pub use cached_event_store::CachedEventStore;
+pub use cached_query_store::CachedQueryStore;
 pub use i_event_dispatcher::IEventDispatcher;
 pub use i_event_store::IEventStore;
 pub use i_query_store::IQueryStore;
 pub use repository::Repository;
 
+mod cached_event_store;
+mod cached_query_store;
 mod i_event_dispatcher;
 mod i_event_store;
 mod i_query_store;